+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 **Version 1.0.0.2008** *(TBD)*
 - Added the non-exempt encryption flag.
+- Added accessibility strings.
 
 **Version 1.0.0.2007** *(November 16, 2018)*
 - Made it so there's no wait, once the alarm's minute is crossed (Happens right at the transition, now. Before, it could wait a few seconds).